Refactors language selection into a modal
Moves the language selection logic into a reusable modal component.
This improves code organization and promotes reusability by
decoupling the language selection from the plugin's main file.
