+++ src/modals/index.ts
@@ -1,0 +1,1 @@
+export { SelectLanguageModal } from './SelectLanguageModal';
