+++ src/plugins/main.ts
@@ -1,10 +1,11 @@
-import { App, Editor, MarkdownView, Modal, Notice, Plugin, Setting } from 'obsidian';
+import { Editor, MarkdownView, Notice, Plugin } from 'obsidian';
 import {
 	FormattedPasterPluginSettings,
 	DEFAULT_SETTINGS,
 	FormattedPasterPluginSettingTab
 } from '../settings';
 import { getClipboardText, insertAtCursor } from '../utilities';
+import { SelectLanguageModal } from '../modals';
 
 export default class FormattedPasterPlugin extends Plugin {
 	settings: FormattedPasterPluginSettings;
@@ -56,7 +57,7 @@ export default class FormattedPasterPlugin extends Plugin {
 					return;
 				}
 
-				new ChosenLanguageModal(this.app, this.settings.code_language, (language) => {
+				new SelectLanguageModal(this.app, this.settings.code_language, (language) => {
 					if (language === undefined) {
 						return;
 					}
@@ -88,49 +89,3 @@ export default class FormattedPasterPlugin extends Plugin {
 	}
 }
 
-export class ChosenLanguageModal extends Modal {
-	private onChoose?: (language?: string) => void;
-	private code_language: Record<string, string>;
-	constructor(app: App, code_language: Record<string, string>, onChoose?: (language?: string) => void) {
-		super(app);
-		this.onChoose = onChoose;
-		this.code_language = code_language;
-	}
-
-	onOpen() {
-		const { contentEl } = this;
-
-		contentEl.addEventListener('keydown', (event) => {
-			if (event.key === 'Escape') {
-				this.onChoose?.(undefined);
-				this.close();
-			}
-
-			if (event.key === 'Enter') {
-				const dropdown = contentEl.querySelector('select');
-				if (dropdown) {
-					const value = (dropdown as HTMLSelectElement).value;
-					console.log('Chosen language: ' + value);
-					this.onChoose?.(value);
-					this.close();
-				}
-			}
-		})
-
-		new Setting(contentEl)
-			.setName('Choose Language')
-			.setDesc('Select a programming language for the code block')
-			.addDropdown(dropdown => {
-				dropdown.addOptions(this.code_language)
-					.setValue('csharp')
-					.onChange((value) => {
-						console.log('Chosen language: ' + value);
-						this.onChoose?.(value);
-						this.close();
-					})
-			})
-	}
-
-	onClose() {
-	}
-}
